+++ tempest_fastapi_sdk/__init__.py
@@ -315,7 +315,7 @@ from tempest_fastapi_sdk.websockets import (
     make_websocket_router,
 )
 
-__version__: str = "0.71.0"
+__version__: str = "0.71.1"
 
 __all__: list[str] = [
     "BASE_COLUMN_ORDER",